xlen.h test-xlen.c platform.h: initial revision
Let's do an LZ engine!  I've got a few half-working snippets of code
scattered over a few directories, some of them going back to 2018,
some are recent.  Now I'm trying to put things together.

The engine is going to be similar to LZ4.  One thing LZ4 got wrong
is how the lengths of long literals and matches are encoded.  When
llen >= 15, or mlen >= 4+15, LZ4 issues an extra byte which encodes
the extra length.  If the extra length is >= 255, LZ4 puts another byte
which encodes (xlen - 255), etc.  From a practical perspective, what
matters most is the first two bytes and what they can do for us (because
very long lengths are rare).  So LZ4 does:

    if (llen >= 15) {
	x = *src++;
	llen += x;
	if (x == 255) {
	    x = *src++;
	    len += x;
	    ...
	}
    }

In this way, two extra bytes can encode only 255+255=510 extra length
(or rather 509, because 510 triggers the third byte).

An alternative to this is vbyte encoding, which uses 7 bits for the
value and the high bit to indicate if there is another byte which
comes with extra MSB bits:

    if (llen >= 15) {
	x = *src++;
	llen += x;
	if (x & 0x80) {
	    x = *src++;
	    len += x << 7;
	    ...
	}
    }

In this way, two bytes can encode 2^(7+7)=16K extra length.  This
is much better because it takes O(log len) to encode len rather than
O(len).  However, lengths in the range [128,255) are at a disadvantage:
they now take another byte to encode and another branch misprediction in
the decoder.

Note that (x & 0x80) can be rewritten as (x >= 128).  Another way to
look at it is that with (x == 255) test,  we have one dedicated value
with special meaning, while with (x >= 128) half of all the values are
special.  We need the middle ground where less than half of the values
are special.  So in "xlen.h" below, the first byte has 16 special values:

    if (llen >= 15) {
	x = *src++;
	llen += x;
	if (x >= 256 - 16) {
	    x = *src++;
	    len += 16 * x;
	    ...
	}
    }

In this way, two bytes can encode up to about 16*256=4K extra length,
but lengths in the range [128,240) are no longer disadvantaged.  This
is an overall better trade-off, because the range [128,240) is more
important than [4K,16K): the extra penalty for the latter range is
barely noticeable (because we make such a big progress by copying so
many literals or such a long match).

I checked that 16 special values on the first byte work best (better
than 8 or 32).  What else do we want?  I believe the lengths that we
encode should be limited.  In LZ4, the lengths are unlimited, which
has consequences:
- you have to run an inner loop (which is inevitable if the encoded
  length takes O(len) bytes);
- on each iteration, LZ4 has to check if there is still enough input;
- on each iteration, LZ4 does NOT check if there is an integer overflow,
  but it should.

Note though that LZ4 actually cannot compress big inputs: in <lz4.h>, we
have LZ4_MAX_INPUT_SIZE 0x7E000000.  The limit seems in part ungrounded,
but in part it has to do with the fact that during compression, the
positions that the match finder stores in the hash table are 32-bit.
The justified limit for such a match finder is about ((4<<30)-(64<<10)),
where (64<<10) is max distance.

So we want to encode lengths slightly less than 4GB.  We want the
encoded length to take at most 5 bytes.  We want 16 special values on
the first byte.  These three requirements pretty much determine the rest
of our construction: the second byte gets 64 special values, the third
128, and the fourth 127 (not 128, to keep max length under 4GB).