xlen.h: use 16-bit last segment (1+1+1+2=5 bytes)
To simplify the code and reduce its size, I choose not to partition the
last 2 bytes in 5-byte encoding.  This affects lengths that are longer
than about 66K - they are now encoded with one extra byte.  But there
is also a plus side: we can extend the range of the second byte, by
reducing the number of special values on the second byte from 64 to 32.
The net result is a win!

 lzbench 1.8 (64-bit Linux)   Assembled by P.Skibinski
 Compressor name         Compress. Decompress. Compr. size  Ratio Filename
 memcpy                   7858 MB/s  7874 MB/s  1227730954 100.00 108 files
-relz4 0.1                  36 MB/s  3589 MB/s   468963290  38.20 108 files
+relz4 0.1                  36 MB/s  3589 MB/s   468961615  38.20 108 files
                                                      ^^^^