change the format to preload the next token even ealier
To load the next token even earlier, we need to change the format.
The traditional format is:

    [T|LLL|MO] [T|LLL|MO] ... [T|LLL]

where square brackets denote a sequence (literals + match); T is the
token, LLL is the literals (a variable number), and MO is the match
offset.  Extra bytes that encode long lengths not shown.  The last
sequence is incomplete.  We now move to the "skewed" format:

    [T] [T|LLL|MO] [T|LLL|MO] ... [T|LLL]

where [T] provides the prime token ("prime" akin to primordial, like
zatravka in Russian, i.e. providing the initial value).  The first byte
in each sequence is the *next* token.  The last sequence also has a
token; for now we just drop it, but it should be possible to reinterpret
it as one of the last literals.

The effect is pretty big, we get 5-7% decompression speed-up!

-relz4 0.1                  34 MB/s  3388 MB/s   468857508  38.19 108 files
+relz4 0.1                  34 MB/s  3615 MB/s   468857616  38.19 108 files