hc.c: better handling of ties in HC_find
This amends the previous change.  We should keep the match with a lower
start (i.e. the one that takes more preceding bytes with src--, ref--);
this reduces the compressed size by 0.021% on my big corpus.  But we
should also prefer match offset >= 128.

Now, which one should take precedence?  Should we give up a lower start
for a higher offset?  It looks like we should.  The difference is
relatively small, but the ratio Δsped/Δsize is favorable.

This redeems most of the compression loss.

-relz4 0.1                  33 MB/s  3898 MB/s   468953423  38.20 108 files
+relz4 0.1                  33 MB/s  3904 MB/s   468861669  38.19 108 files
 relz4 0.1                  34 MB/s  3782 MB/s   468857616  38.19 108 files (old)