hc.c: preload htab[] value early
 lz4hc 1.9.3 -4           59.7 MB/s  2741 MB/s   474572474  38.65 108 files
-relz4 0.1 -4             55.4 MB/s  3867 MB/s   474310884  38.63 108 files
+relz4 0.1 -4             57.9 MB/s  3868 MB/s   474310884  38.63 108 files

At level 4, I'm actually able to beat lz4hc in size!  This is probably
because my parser tries harder to go back and take preceding bytes.
(In lz4hc, when LZ4HC_InsertAndFindBestMatch is called in a loop, there
is no going back.  If no match is found, that position is forever lost
to a literal.  But with more searches, this is less likely.)