hc.c: double the number of HC_find0 searches
HC_find0 is faster than HC_find because it does not backtrack, it has
clever probing, and the probing can reduce the number of HC_count
iterations.  Let's give HC_find0 a chance with 2x iterations!  This is
not necessarily slower: if we are more likely to find a match with
HC_find0, we are less likely to proceed with a more expensive loop.
(It's a bit slower in the end, but it's close, and the compression gain
is arguably worth it.  We can rebalance and fine-tune the exact number
of iterations later.)

 lz4hc 1.9.3 -3           70.3 MB/s  2703 MB/s   480589267  39.14 108 files
-relz4 0.1 -3             74.1 MB/s  3800 MB/s   480014533  39.10 108 files
+relz4 0.1 -3             71.6 MB/s  3829 MB/s   478584618  38.98 108 files

 lz4hc 1.9.3 -7           38.3 MB/s  2781 MB/s   468133763  38.13 108 files
-relz4 0.1 -7             36.3 MB/s  3862 MB/s   468714261  38.18 108 files
+relz4 0.1 -7             34.9 MB/s  3876 MB/s   468620241  38.17 108 files