hc.c: probe for a longer match in HC_find
In HC_find, we cannot always probe for a longer match, because there
are preceding bytes, and we can get a longer match as a result of
backtracking.  We do not want to preclude this possibility.

          probe
          vvvv
    -----====----->
    ^    ^
    src0 src1

The mandatory 4-byte match is shown with '===='.  A meaningful probe
for a longer match must go at least 1 byte past that region (shown with
"vvvv").  This is only possible if (src0 + bestmlen > src1 + 4).
In other words, to probe for a longer match, we must assume that it
may start at src0.  If the region (src1 - src0) is too long while the
existing match is all too short, no such probe exists.

-relz4 0.1 -7        34.8 MB/s  3874 MB/s   468620241  38.17 108 files  de3d3d60
+relz4 0.1 -7        35.2 MB/s  3896 MB/s   468620241  38.17 108 files  de3d3d60