Revert "hc.c: probe for a longer match in HC_find"
Microbenchmarking shows no improvement, and there is an explanation.
Suppose we've found a long match at src0, mlen=16.  We try to find an
overlapping match at src1=src0+mlen-3.

                  probe
                  vvvv
    -------------====----->
    ^            ^
    src0         src1

We have minprobe=src1-src0+4=13+4=17, and the precondition to apply
the probe is bestmlen > 17.  This means that the probe cannot be applied
*before* you try to find a longer match, only *after* you have already
found it.  The probe is only effective for "even longer" matches.