ht.c: backtrack up to 3 bytes in HT_find1
Two would be enough, but three seem to run a tad bit faster (not sure why).

With probing, you cannot backtrack in a loop.  The worst case is that,
in a previous iteration, the match length is already the best possible
and hits last5.  In the current iteration, if you backtrack too many
bytes, you can further increase bestmlen.  Then in the next iteration,
the probe might hit past the end of input.

-relz4 0.1 -2        92.0 MB/s  3827 MB/s   482082606  39.27 108 files  c88c6c5e
+relz4 0.1 -2        92.6 MB/s  3831 MB/s   482053512  39.26 108 files  e943f03c