ht.c: specialize HT_find1
(This is not a proper "instantiation" or "specialization" because it
changes the output.)  The idea is that backtracking is rare, and when
we do backtrack, we usually go back only by one byte.  A clever
compromise is to proceed as if no backtracking would happen.  However,
when it does happen, we try to take advantage of it.

So to find a new match with pending literals, we can use a variant of
find0.  This seems to fit nicely with the other parts of the algorithm:
1) NICEOFF logic still applies; 2) it is more likely that we select
a lower start, although we do not check for it explicitly.

-relz4 0.1 -2        90.9 MB/s  3833 MB/s   481961521  39.26 108 files  442adf43
+relz4 0.1 -2        93.0 MB/s  3835 MB/s   482082606  39.27 108 files  c88c6c5e (find1)
 relz4 0.1 -2        94.0 MB/s  3834 MB/s   482609608  39.31 108 files  12725e4a (find0)

By trying to backtrack one byte only, we manage to redeem most of the
compression loss.