ht.c: should we extend match length 19 though?
If we leave match length 19 as it is, putseq writes match length 18
and we have 1 pending literal.  If we extend it to 20+, we get an
additional length byte and a branch misprediction.  But there is also
a chance to shorten the next match, which may change how we deal with
it later.  Let's see how many long matches we get.

0) long matches not extended:    7101822 long matches out of 79456139 total
1) extend all long matches:      7111258 ( +9436)     out of 79439356 (-16783)
2) do not extend mlen 19 though: 7091309 (-19949)     out of 79442460 ( +3104)

Decompression speed (999 iterations, takes a few minutes):
no significant difference.

0) 3837 MB/s   482053512  39.26 108 files  e943f03c
1) 3836 MB/s   482004368  39.26 108 files  43404f23
2) 3837 MB/s   482009802  39.26 108 files  8ba78652

Let's say on a 3 GHz CPU, the decoding speed is 4 GB/s (so it
takes 3 cycles to decompress 4 bytes).  An average token comes with
1227730954/7091309=15.45 bytes, let that be 16.  So a token takes
12 cycles to process - on average, including long matches.
Short matches are cheaper.  With long matches, branch misprediction
penalty is 15 cycles.  This means we should generally opt for short
matches, e.g. we should trade one 20+ match for two sub-19 matches,
other things being equal.  So the change looks favorable after all:
we have a lot less long matches, and not that much more total matches.