ht.c: bump minimum (mstart2 - mstart) distance 3->4
When M2 is only 3 characters away from M, we have a tie: we can either
encode two sequences M and M2 (with M borrowing an additional byte for
its 4-byte match from M2), or encode M2 with 3 preceding literals.
When it is truly a tie, we are better off with the latter option,
because a single sequence is faster to process on decoding than two
sequences.  However, it is not always a true tie.  For one, the number
of literals preceding M can be less than OPTLLEN, but if we add 3 more,
it can become more than OPTLLEN (so it takes an extra llen byte and a
branch misprediction on decoding).  Similarly, if we dismiss M+M2 too
soon, we miss the opportunity of moving the boundary between M and M2 to
the right, which makes M2 shorter and can save us an extra mlen byte.

But let's try it anyway.

-relz4 0.1 -2             93.5 MB/s  3834 MB/s   481984498  39.26 108 files  c7625b1f
+relz4 0.1 -2             94.3 MB/s  3860 MB/s   482408741  39.29 108 files  05d09303

-7091196 long matches out of 79453495 total
+7217640 (+126444)           77708306 (-1745189)

We've lost 424K bytes, but have 1.7M fewer tokens, with a measurable
improvement in the decoding speed.  If we revert the change, we'll have
to admit than one extra token saves us only 0.24 bytes.  That would be
a poor deal.  But loosing 424K isn't a fun either.  We must try and
think of something smarter than that. :)