ht.c: more precise condition for merging M+M2 into L+M2
-482259341 compressed bytes
+482273635 compressed bytes

So we can only save 16K by tracking OPTLLEN boundary and about 120K by
tracking OPTMLEN boundary, out of 424K.  Crossing the boundary is not
the only way of making compression less efficient.  If we don't shorten
M2, we're less likely to find a longer match on the next iteration.

-7105392/78063879 tokens
+7125693/77842085 tokens