ht.c: probes for HT_find
Can we make a probe if a match can be expanded both forward and
backward?  It seemed hopeless at some point.  However, once we realize
that we always attempt to descend backward, we can move this check to
the very top.  So that if we don't need to descend, we can probe in a
straightforward way, much like in HT_find0.  If we do descend though,
we can probe after the descent.  The probe is less efficient, because
the effort to descend may turn out to be wasted.  However, we can extend
this probe to favor lower match start.

Zen2 2.4 GHz:
-relz4 0.1 -2  83.4 MB/s  3597 MB/s  482223668  39.28  108 files  e3929400
+relz4 0.1 -2  91.3 MB/s  3598 MB/s  482223668  39.28  108 files  e3929400

The checksum is the same!

There's a problem though I'm not yet sure how to deal with: the probe
may overshoot past the end of input.  Note that, in HT_find0, the probe
starts with bestmlen=3, and we don't extend the match past last5.

In HT_find, we have bestmlen of the match that we found at some lower
start, but the new start is

	// search for a longer overlapping match
	src = mstart + mlen - 2 - (mlen > 4);

So the probe may hit past the end of input immediately!