ht.c: improve *pmstart initialization in HT_find
pmoff and pmstart must be initialized in such a way that the first
iteration looks for a longer match (i.e. we aren't interested in the
same length).  They are initialized with fake values until we find a
longer match, at which point they are reset to real values (and if we
don't find a longer match, fake values won't be used).

    *pmoff = NICEOFF;
    *pmstart = src1;

So we normally look for a longer match, unless we've descended to
a lower start (in which case we'll take the new match with the
same length, just because the start is lower).  So we probe for
a longer match only if (src >= *pmstart) aka "have not descended
low enough".  However, since we extend the match backward by repeatedly
doing src--, the condition (src >= *pmstart) will not hold on the first
iteration.

    while (src > src0 && ref > ht->base && src[-1] == ref[-1])
	src--, ref--, mlen++;
    // probe for a longer match, unless the offset is small,
    // or unless we've descended to a smaller start
    uint32_t probe = bestmlen + (*pmoff >= NICEOFF && src >= *pmstart);
    if (load32(ref + probe - 4) != load32(src + probe - 4))
	continue;

Instead, *pmstart should be initialized with src0 (or with NULL,
but (src >= NULL) would be a questionable comparison from the
perspective of the C standard).

No change in the output, but the probe should be slightly more efficient
(the "continue" statement gets executed more often).