ht.c: unlikely to extend a match backward in HT_find1
Since HT_find1 is called in a loop until a match is found, the only way
it can extend a match backward is because we didn't try hard enough on
the previous iteration (in other words, because the number of positions
in the match structure limited to 4).  The probability of extending
a match backward is about 0.2%.  (This is very different to what happens
in HT_find, where we deliberately probe for the segment in the middle
of a match.  The probability of extending a match backward in HT_find
is about 30%).