ht.c: revert to a simpler loop in HT_find1
Opportunistic backtracing (this commit):
relz4 0.1 -2   95.88 MB/s  3570 MB/s  482223196  39.28  108 files  c1ebf6c2

Accurate backtracing (previous commit):
relz4 0.1 -2   88.64 MB/s  3573 MB/s  482132837  39.27  108 files  595d7713

Don't look back into the strange face of love (find1 equivalent to find0):
relz4 0.1 -2   95.72 MB/s  3570 MB/s  482767086  39.32  108 files  e61178c5

So despite all my efforts the previous version of find1 was a bit too
slow.  I want HT to be more practical: it must not waste CPU cycles in
a way that is unlikely to achieve anything.  With no backtracking at
all, we lose 482767086-482132837=634249 bytes, or about 0.13%.  That's
not much, but not negligible either.  With opportunistic backtracking,
which is much faster than accurate backtracing, we redeem 543890 of
those bytes, while losing only 482223196-482132837=90359 bytes.

So why is opportunistic backtracing much faster?  It might seem that
we attempt to backtrack either way.  The answer is that the probe for
a longer match must be executed ASAP.  This is because find1 runs in
a loop which tries to find a match, while in fact there might be no
matches at all.  So probe first, probe ASAP, to discard unpromising
positions!